+++ src/Models/Queries/HotspotAssetQuery.php
@@ -15,7 +15,7 @@ class HotspotAssetQuery extends I18nActiveQuery
 {
     public function selectSiteAttributes(): self
     {
-        return $this->addSelect($this->prefixColumns(array_diff($this->getModelInstance()->attributes(), [
+        return $this->addSelect($this->prefixColumns(array_diff($this->getModelInstance()->getColumnAttributes(), [
             'updated_by_user_id',
             'created_at',
         ])));
@@ -26,7 +26,7 @@ class HotspotAssetQuery extends I18nActiveQuery
         return $this->with([
             'file' => function (FileQuery $query): void {
                 $query->selectSiteAttributes()
-                    ->replaceI18nAttributes()
+                    ->withTranslations()
                     ->withFolder();
             }
         ]);