+++ src/Models/Queries/HotspotQuery.php
@@ -14,7 +14,7 @@ class HotspotQuery extends I18nActiveQuery
 {
     public function selectSiteAttributes(): static
     {
-        return $this->addSelect($this->prefixColumns(array_diff($this->getModelInstance()->attributes(), [
+        return $this->addSelect($this->prefixColumns(array_diff($this->getModelInstance()->getColumnAttributes(), [
             'updated_by_user_id',
             'created_at',
         ])));